// src/enums/NormalFishType.ts
export enum NormalFishType {
  // ARR
  MERLTHOR_GOBY = 4869,
  OCEAN_CLOUD = 4872,
  HARBOR_HERRING = 4874,
  OGRE_BARRACUDA = 4888,
  FULLMOON_SARDINE = 4898,
  WAHOO = 4904,
  LITTLE_THALAOS = 4913,
  GIANT_SQUID = 4919,
  STRIPED_GOBY = 4927,
  ABALATHIAN_SMELT = 4937,
  ALA_MHIGAN_FIGHTING_FISH = 4942,
  COPPERFISH = 4948,
  SILVERFISH = 4978,
  COMMON_SCULPIN = 4995,
  ASSASSIN_BETTA = 5002,
  BOXING_PLECO = 5008,
  GOLDFISH = 5011,
  JUNGLE_CATFISH = 5031,
  STORM_RIDER = 5035,
  CLOUD_CUTTER = 5040,
  LAMP_MARIMO = 5544,

  // HW
  ICE_FAERIE = 12715,
  BLUECLAW_SHRIMP = 12722,
  GLACIER_CORE = 12724,
  BULLFROG = 12730,
  SKY_FAERIE = 12753,
  GRANITE_CRAB = 12754,
  HEDGEMOLE_CRICKET = 12757,
  PLATINUM_FISH = 12776,
  AETHER_EYE = 12777,
  SWEETFISH = 12780,
  FOSSILTONGUE = 12800,
  RUDDERFISH = 12805,
  SCORPIONFLY = 12810,
  FUNCTIONAL_PROTO_HROPKEN = 13727,

  // STB
  ROCK_SALTFISH = 20025,
  SCULPTOR = 20040,
  GYR_ABANIAN_TROUT = 20056,
  BALLOON_FROG = 20064,
  RUBY_SHRIMP = 20112,
  ZAGAS_KHAAL = 20127,
  FIRELIGHT_GOLDFISH = 23056,
  INDIGO_PRISMFISH = 24203,
  GREEN_PRISMFISH = 24204,
}