// src/enums/BaitType.ts
export enum BaitType {
  // ARR
  LUGWORM = 2585,
  MOTH_PUPA = 2586,
  PILL_BUG = 2587,
  CRAYFISH_BALL = 2588,
  GOBY_BALL = 2589,
  BLOODWORM = 2590,
  RAT_TAIL = 2591,
  MIDGE_BASKET = 2592,
  BUTTERWORM = 2594,
  BASS_BALL = 2595,
  SPOON_WORM = 2596,
  SYRPHID_BASKET = 2597,
  HERRING_BALL = 2598,
  HONEY_WORM = 2599,
  SAND_LEECH = 2600,
  STEM_BORER = 2601,
  GLOWWORM = 2603,
  HOVERWORM = 2605,
  NORTHERN_KRILL = 2606,
  CADDISFLY_LARVA = 2607,
  CHOCOBO_FLY = 2610,
  FLOATING_MINNOW = 2611,
  SHRIMP_CAGE_FEEDER = 2613,
  CROW_FLY = 2614,
  STEEL_JIG = 2616,
  SINKING_MINNOW = 2617,
  SNURBLE_FLY = 2618,
  HEAVY_STEEL_JIG = 2619,
  SPINNER = 2620,
  WILDFOWL_FLY = 2623,
  TOPWATER_FROG = 2624,
  RAINBOW_SPOON_LURE = 2625,
  SPINNERBAIT = 2626,
  YUMIZUNO = 2628,

  // HW
  STONEFLY_NYMPH = 12704,
  PURSE_WEB_SPIDER = 12705,
  GOBLIN_JIG = 12706,
  BLADED_STEEL_JIG = 12707,
  RED_BALLOON = 12708,
  MAGMA_WORM = 12709,
  FIEND_WORM = 12710,
  BRUTE_LEECH = 12711,
  GIANT_CRANE_FLY = 12712,
  METAL_SPINNER = 28634,

  // STB
  SALMON_ROE = 20613,
  NIGHTCRAWLER = 20614,
  MIDGE_LARVA = 20615,
  SILKWORM = 20616,
  LIVE_SHRIMP = 20617,
  BREAM_LURE = 20618,
  SUSPENDING_MINNOW = 20619,
  STONEFLY_LARVA = 20675,
  BLUE_BOBBIT = 20676,
}